+++ src/assembler.rs
@@ -568,13 +568,12 @@ fn parse_parts(source: &str) -> Result<(Module, Vec<FieldFixup>), Fault> {
                         "sizeof" | "alignof" | "heap.alloc" | "ptr.null" | "ptr.cast"
                         | "ptr.fromint" | "ldobj" | "stobj" | "cpobj" | "initobj" | "box"
                         | "value.pack" | "value.is" | "value.unpack" | "ldtoken" | "castclass"
-                        | "interface.borrow" | "newarr" | "array.new" | "array.alloc"
-                        | "array.reserve" | "array.create" | "ldelem" | "stelem" | "ldelema" => {
-                            Some(
-                                serde_json::to_value(parse_type(rest)?)
-                                    .map_err(|e| Fault::new(e.to_string()))?,
-                            )
-                        }
+                        | "isinst" | "interface.borrow" | "newarr" | "array.new"
+                        | "array.alloc" | "array.reserve" | "array.create" | "ldelem"
+                        | "stelem" | "ldelema" => Some(
+                            serde_json::to_value(parse_type(rest)?)
+                                .map_err(|e| Fault::new(e.to_string()))?,
+                        ),
                         _ => {
                             if !rest.is_empty() {
                                 return Err(Fault::new("unexpected instruction operand"));